Move timeline in separate file
